requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
google-generativeai==0.8.5
mcp==1.1.0
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.36
